+++ imeditor/editor/tools.py
@@ -13,13 +13,14 @@ def get_infos(img, filename):
     """Fetch informations about an image"""
     img_infos = {}
     # Basic infos
+    img_infos['name'] = path.basename(filename)
     img_infos['mode'] = img.mode
     img_infos['size'] = '{} x {} pixels'.format(str(img.width), str(img.height))
     # Infos available only if the image is save on the disk
     if path.isfile(filename):
         img_stat = stat(filename)
-        img_infos['weight'] = '{}ko ({}o)'.format(str(round(img_stat.st_size / 1000, 2)), str(round(img_stat.st_size, 2)))
-        img_infos['path'] = filename
-        img_infos['last_access'] = datetime.datetime.fromtimestamp(img_stat.st_atime).strftime('%d/%m/%Y %Hh%M')
+        img_infos['weight'] = '{}ko ({}o)'.format(str(round(img_stat.st_size / 1000, 2)),
+                                                    str(round(img_stat.st_size, 2)))
+        img_infos['folder'] = path.dirname(filename)
         img_infos['last_change'] = datetime.datetime.fromtimestamp(img_stat.st_mtime).strftime('%d/%m/%Y %Hh%M')
     return img_infos